Use box-sizing to apply Choice Card border
Using the border property to change the style of the border on hover
causes the layout to jump if the Choice Cards are < 150px wide. By
using `box-shadow: inset`, the border style is always applied within
the bounds of the Choice Card's label element, so the width of the
element is not changed.